# Conflicts:
#	renderlib/FileReader.cpp

+++ renderlib/Histogram.cpp
@@ -153,7 +153,8 @@ float* Histogram::generate_auto2(float& window, float& level, size_t length) {
 			break;
 		}
 	}
-	for (size_t i = nbins - 1; i >= 0; --i) {
+    // need a signed loop control variable here to test for >= 0.
+	for (int i = (int)nbins - 1; i >= 0; --i) {
 		if (_bins[i] > threshold && _bins[i] <= limit) {
 			hmax = i;
 			break;
